pyproject.toml
[build-system]
requires = ['flit_core >=3.4,<4']
build-backend = 'flit_core.buildapi'

[project]
name = 'aiida-abinit'
dynamic = ['description', 'version']
authors = [{name = 'Samuel Ponce', email = 'samuel.pon@gmail.com'}]
readme = 'README.md'
license = {file = 'LICENSE.txt'}
classifiers = [
    'Development Status :: 4 - Beta',
    'Framework :: AiiDA',
    'License :: OSI Approved :: MIT License',
    'Operating System :: POSIX :: Linux',
    'Operating System :: MacOS :: MacOS X',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
keywords = ['aiida', 'abinit']
requires-python = '>=3.10'
dependencies = [
    'aiida_core>=2.3,<3',
    'aiida-pseudo>=1.0',
    'abipy>=0.9.4',
    'packaging',
    'pymatgen',
    'numpy',
    'importlib_resources'
]

[project.urls]
Home = 'https://github.com/sponce24/aiida-abinit'
Source = 'https://github.com/sponce24/aiida-abinit'

[project.optional-dependencies]
docs = [
    'myst-nb~=1.0',
    'jupytext>=1.11.2,<1.15.0',
    'sphinx~=6.2.1',
    'sphinx-copybutton~=0.5.2',
    'sphinx-book-theme~=1.0.1',
    'sphinx-click~=4.4.0',
    'sphinx-design~=0.4.1',
    'sphinxcontrib-details-directive~=0.1.0',
    'sphinx-autoapi~=3.0.0',
    'myst-parser~=3.0.0',
    'sphinx-togglebutton',
]
pre-commit = [
    'pre-commit~=2.17',
    'pylint~=2.15.10',
    'pylint-aiida~=0.1.1',
    'toml'
]
tests = [
    'pgtest~=1.3',
    'pytest>=7.0',
    'coverage[toml]',
    'pytest-cov',
    'pytest-regressions~=2.3',
    'pytest-timeout',
]

[project.entry-points.'aiida.calculations']
'abinit' = 'aiida_abinit.calculations:AbinitCalculation'
'abinit.mrgddb' = 'aiida_abinit.calculations:MrgddbCalculation'
'abinit.anaddb' = 'aiida_abinit.calculations:AnaddbCalculation'

[project.entry-points.'aiida.parsers']
'abinit' = 'aiida_abinit.parsers:AbinitParser'
'abinit.mrgddb' = 'aiida_abinit.parsers:MrgddbParser'
'abinit.anaddb' = 'aiida_abinit.parsers:AnaddbParser'

[project.entry-points.'aiida.workflows']
'abinit.base' = 'aiida_abinit.workflows.base:AbinitBaseWorkChain'

[tool.flit.module]
name = 'aiida_abinit'

[tool.flit.sdist]
exclude = [
    '.github/',
    'docs/',
    'tests/',
]

[tool.flynt]
line-length = 120
fail-on-change = true

[tool.isort]
force_sort_within_sections = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3

[tool.pydocstyle]
ignore = [
    'D104',
    'D202',
    'D203',
    'D213'
]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.tool-check]
generated-members = 'self.exit_codes.*'

[tool.pylint.messages_control]
disable = [
    'duplicate-code',
    'locally-disabled',
    'logging-format-interpolation',
    'inconsistent-return-statements',
    'import-outside-toplevel',
    'no-else-raise',
    'too-many-arguments',
    'too-many-ancestors',
    'too-many-branches',
    'too-many-locals',
]

[tool.pylint.basic]
good-names = [
    'i',
    'j',
    'k',
    'SsspFamily',
    'StructureData',
    'UpfData',
    'Psp8Data',
    'JthXmlData',
    'AbinitCalculation',
]

[tool.pytest.ini_options]
filterwarnings = [
    'ignore:Creating AiiDA configuration folder.*:UserWarning',
    'ignore::DeprecationWarning:frozendict:',
    'ignore::DeprecationWarning:pkg_resources:',
    'ignore::DeprecationWarning:reentry:',
    'ignore::DeprecationWarning:sqlalchemy_utils:',
]
minversion = '7.0'
testpaths = [
    'tests',
]

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = 'google'
coalesce_brackets = true
column_limit = 120
dedent_closing_brackets = true
indent_dictionary_value = false
split_arguments_when_comma_terminated = true

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py38

[testenv]
usedevelop=True

[testenv:py{38,39}]
extras = tests
commands = pytest {posargs}

[testenv:py38-pre-commit]
description = Run the pre-commit checks
extras =
    tests
    pre-commit
commands = pre-commit run {posargs}

[testenv:py38-docs-{clean,update}]
description =
    clean: Build the documentation (remove any existing build)
    update: Build the documentation (modify any existing build)
extras = docs
changedir = docs
whitelist_externals = make
commands =
    clean: make clean
    make

"""
